requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
antiorm==1.2.1
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.21.0
asyncpgsa==0.26.3
attrs==20.3.0
Babel==2.9.0
chardet==3.0.4
db==0.1.1
db-sqlite3==0.0.1
gino==1.0.1
idna==2.10
multidict==5.0.2
python-dotenv==0.15.0
pytz==2020.4
typing-extensions==3.7.4.3
wincertstore==0.2
yarl==1.6.3
